ReservasMedicasWeb/BackEnd/requirements.txt
asgiref==3.8.1
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2 
mysqlclient==2.2.4
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
pip==24.0
pytz==2023.3
coreapi==2.3.3
